One more change to building Kubernetes Client I missed
